refactor: improving code readability

+++ Controllers/HomeController.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Todo.Data;
+using Todo.Migrations;
 using Todo.Models;
 
 namespace Todo.Controllers
@@ -8,27 +9,29 @@ namespace Todo.Controllers
     public class HomeController : ControllerBase
     {
         [HttpGet("/")]
-        public List<TodoModel> Get([FromServices] AppDbContext ctx)
-        {
-            return ctx.Todos.ToList();
-        }
+        public IActionResult Get([FromServices] AppDbContext ctx)
+            => Ok(ctx.Todos.ToList());
 
         [HttpGet("/{id:int}")]
-        public TodoModel Get([FromRoute] int id, [FromServices] AppDbContext ctx)
+        public IActionResult GetById([FromRoute] int id, [FromServices] AppDbContext ctx)
         {
-            return ctx.Todos.FirstOrDefault(x => x.Id == id);
+            var todos = ctx.Todos.FirstOrDefault(x => x.Id == id);
+            if (todos == null)
+                return NotFound();
+
+            return Ok(todos);
         }
 
         [HttpPost("/")]
-        public TodoModel Post([FromBody] TodoModel todo, [FromServices] AppDbContext ctx)
+        public IActionResult Post([FromBody] TodoModel todo, [FromServices] AppDbContext ctx)
         {
             ctx.Todos.Add(todo);
             ctx.SaveChanges();
-            return todo;
+            return Create($"/{todo.Id}", todo);
         }
 
         [HttpPut("/{id:int}")]
-        public TodoModel Put(
+        public IActionResult Put(
             [FromRoute] int id,
             [FromBody] TodoModel todo,
             [FromServices] AppDbContext ctx
@@ -37,24 +40,28 @@ namespace Todo.Controllers
             var model = ctx.Todos.FirstOrDefault(x => x.Id == id);
             if (model == null)
             {
-                return todo;
+                return NotFound();
             }
 
             model.Title = todo.Title;
             model.Done = todo.Done;
             ctx.Todos.Update(model);
             ctx.SaveChanges();
-            return model;
+            return Ok(model);
         }
 
         [HttpDelete("/{id:int}")]
-        public TodoModel Put([FromRoute] int id, [FromServices] AppDbContext ctx)
+        public IActionResult Put([FromRoute] int id, [FromServices] AppDbContext ctx)
         {
             var model = ctx.Todos.FirstOrDefault(x => x.Id == id);
+            if (model == null)
+            {
+                return NotFound();
+            }
 
             ctx.Todos.Remove(model);
             ctx.SaveChanges();
-            return model;
+            return Ok(model);
         }
     }
 }
